requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.16
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
filelock==3.10.7
idna==3.4
oauthlib==3.2.2
platformdirs==3.2.0
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
Pillow-11.0.0
requests==2.30.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
urllib3==2.0.2
virtualenv==20.21.0
gunicorn==20.1.0

